+++ aasrp/tests/test_model_srplink.py
@@ -194,7 +194,7 @@ class TestSrpLink(BaseTestCase):
 
             srp_link = SrpLink()
 
-            result = srp_link.total_requests
+            result = srp_link.total_requests_count
 
             self.assertEqual(result, 5)
             mock_rels.return_value.count.assert_called_once()
@@ -214,7 +214,7 @@ class TestSrpLink(BaseTestCase):
 
             srp_link = SrpLink()
 
-            result = srp_link.total_requests
+            result = srp_link.total_requests_count
 
             self.assertEqual(result, 0)
             mock_rels.return_value.count.assert_called_once()
@@ -234,7 +234,7 @@ class TestSrpLink(BaseTestCase):
 
             srp_link = SrpLink()
 
-            result = srp_link.pending_requests
+            result = srp_link.pending_requests_count
 
             self.assertEqual(result, 3)
             mock_rels.return_value.filter.assert_called_with(
@@ -256,7 +256,7 @@ class TestSrpLink(BaseTestCase):
 
             srp_link = SrpLink()
 
-            result = srp_link.pending_requests
+            result = srp_link.pending_requests_count
 
             self.assertEqual(result, 0)
             mock_rels.return_value.filter.assert_called_with(
@@ -278,7 +278,7 @@ class TestSrpLink(BaseTestCase):
 
             srp_link = SrpLink()
 
-            result = srp_link.approved_requests
+            result = srp_link.approved_requests_count
 
             self.assertEqual(result, 4)
             mock_rels.return_value.filter.assert_called_with(
@@ -300,7 +300,7 @@ class TestSrpLink(BaseTestCase):
 
             srp_link = SrpLink()
 
-            result = srp_link.approved_requests
+            result = srp_link.approved_requests_count
 
             self.assertEqual(result, 0)
             mock_rels.return_value.filter.assert_called_with(
@@ -322,7 +322,7 @@ class TestSrpLink(BaseTestCase):
 
             srp_link = SrpLink()
 
-            result = srp_link.rejected_requests
+            result = srp_link.rejected_requests_count
 
             self.assertEqual(result, 2)
             mock_rels.return_value.filter.assert_called_with(
@@ -344,7 +344,7 @@ class TestSrpLink(BaseTestCase):
 
             srp_link = SrpLink()
 
-            result = srp_link.rejected_requests
+            result = srp_link.rejected_requests_count
 
             self.assertEqual(result, 0)
             mock_rels.return_value.filter.assert_called_with(